refactor(model): extract shared container types so the full inspect profile crosses the DI seam
Add internal/model (behavior-free, stdlib-only) holding the rich Inspect,
Config, HostConfig, Mount, PortBinding, DeviceMapping, RestartPolicy types
incl. all security-relevant fields (User, Cap*, Privileged, SecurityOpt,
ReadonlyRootfs, NetworkMode, Devices). dockercli now uses model.* in its Docker
interface and mapping helpers, dropping its duplicate ContainerInspect/Config/
HostConfig. Keeps dependency direction clean: dockercli -> model.
